Worked on blog editor, and codebase improvements. - Broke out footer to component. - Broke out nav to component. - Broke out blogEditor to component to DRY out the code a bit. - Renamed sessionButton to something more fitting. - Made drafts dropdown update on click.
