+++ components/sessionNavSection.jsx
@@ -1,4 +1,3 @@
-import Link from "next/link";
 import { Button } from "reactstrap"
 import { useSession, signIn, signOut } from "next-auth/react"
 import { useEffect, useState } from "react"
@@ -9,25 +8,32 @@ import {
   DropdownItem,
 } from 'reactstrap';
 
-export default function SessionButton() {
+export default function SessionNavSection(props) {
   const [drafts, setDrafts] = useState([])
   const { data: session } = useSession()
-  useEffect(() => {
+
+  const updateDrafts = () => {
     fetch(`/api/posts/drafts`)
-      .then(data =>
-        data.json().then(data => {
-          setDrafts(data)
-        })
-      )
+    .then(data =>
+      data.json().then(data => {
+        setDrafts(data)
+      })
+    )
+  }
+
+  useEffect(() => {
+    updateDrafts()
   }, [session])
+  
   if (session) {
     return (
       <>
-        {drafts && <DraftsDropdown  drafts={drafts} />}
+        {drafts && <DraftsDropdown  drafts={drafts} updateDrafts={updateDrafts} />}
         <Button color="danger" onClick={() => signOut()}>Sign out</Button>
       </>
     )
   }
+  
   return (
     <>
       <p className="color-two mr-4">
@@ -38,11 +44,12 @@ export default function SessionButton() {
   )
 }
 
-function DraftsDropdown({ drafts }) {
+function DraftsDropdown({ drafts, updateDrafts }) {
   return <UncontrolledDropdown group className="mr-4 bg-primary">
     <DropdownToggle
       caret
       color="primary"
+      onClick={updateDrafts}
     >My Drafts</DropdownToggle>
     <DropdownMenu>
       <DropdownItem href={"/blog/new"}>
