+++ pages/api/posts/drafts.js
@@ -33,6 +33,6 @@ export default function handler(req, res) {
       })
   })
 
-      break;
+      // break;
   }
 }
