+++ styles/blog.module.scss
@@ -7,3 +7,11 @@
   top: 1rem;
   right: 1rem;
 }
+
+.blogContent {
+  display: flex;
+  flex-direction: column;
+  pre {
+    color: white;
+  }
+}
